switched to static-files crate for build phase
