+++ Cargo.toml
@@ -18,21 +18,16 @@ version = "3.0.5"
 
 [features]
 default = [ "change-detection" ]
+change-detection = [ "static-files/change-detection" ]
 
 [dependencies]
 actix-web = { version = "3.0", default-features = false }
 actix-service = { version = "1.0.6", default-features = false }
-change-detection = { version = "1.1", optional = true }
 derive_more = "0.99.5"
 futures = { version = "0.3", default-features = false }
-mime_guess = "2.0"
-path-slash = "0.1"
+static-files = { version = "0.1", default-features = false }
 
 [build-dependencies]
-actix-web = { version = "3.0", default-features = false }
-actix-service = { version = "1.0.6", default-features = false }
-change-detection = { version = "1.1", optional = true }
-derive_more = "0.99.5"
-futures = { version = "0.3", default-features = false }
-mime_guess = "2.0"
-path-slash = "0.1"
+# mime_guess = "2.0"
+# path-slash = "0.1"
+static-files = { version = "0.1", default-features = false }
