+++ build.rs
@@ -1,11 +1,16 @@
 #![doc(html_no_source)]
-include!("src/impl.rs");
-fn main() {
-    resource_dir("./tests").build().unwrap();
+use static_files::{
+    resource::{generate_resources_mapping, resource_dir},
+    sets,
+};
+use std::{env, io, path::Path};
+
+fn main() -> io::Result<()> {
+    resource_dir("./tests").build()?;
 
     let out_dir = env::var("OUT_DIR").unwrap();
     let generated_filename = Path::new(&out_dir).join("generated_mapping.rs");
-    generate_resources_mapping("./tests", None, generated_filename).unwrap();
+    generate_resources_mapping("./tests", None, generated_filename)?;
 
     sets::generate_resources_sets(
         "./tests",
@@ -14,6 +19,7 @@ fn main() {
         "sets",
         "generate",
         &mut sets::SplitByCount::new(2),
-    )
-    .unwrap();
+    )?;
+
+    Ok(())
 }
