+++ src/deps.rs
@@ -1,0 +1,1 @@
+pub use static_files::*;
